Funcion for checking victory lines "completed"
I am reasonably confident it works... but it relies on some assumptions;
in particular that there is no current winner.
Also kinda special-casey and reliant on a 4x4x4 board.
Works by checking lines relevant to the most recent move.
Multiplies each square in the line succsesively;
ending with 1 means noughts victory, 16 means crosses victory.
A bit of testing code added to worker.c but... need a better way to
test it. Ideas?