Added some functions to glue together some very basic functionality.
The efficiency gains from only checking victory in the latest state are pretty much ignored, just to test if the core code works. With depth 2, it looks like it blocks all victories, but I haven't fully looked at the board yet to see if anyone is making clearly wrong decisions.